+++ omega_gamecode/code/game/bg_public.h
@@ -49,6 +49,8 @@ Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA  02110-1301  USA
 
 #define LIGHTNING_RANGE 768
 
+#define RAIL_RELOAD_FAST 800
+
 #define SCORE_NOT_PRESENT -9999 // for the CS_SCORES[12] when only one player is present
 
 #define VOTE_TIME 30000 // 30 seconds before vote times out