+++ CHANGELOG.md
@@ -148,3 +148,4 @@
 - better `g_guidedRockets` (don't shoot another rocket if your holding attack)
 - don't force inactive players to go spectator now it's defined by `g_forcespectator`
 - add mutliple explicit cvars (`pmove_doublejump`, `g_spectateOnDeath`, `g_noInvisWalls`)
+- add `g_railgunImpressiveReward` which double the reload speed of the railgun after an "Impressive" medal